Add the ability to weight classes
This was meant to make predicting minority classes easier.
However, the reality is that they're largely useless without also adding a minimum node size, due to nodes reducing their Gini down to a 0.0 (a single class per node). This means that even weights cannot help, because no matter how much you weight a class, if it's nonexistent in the node it's not getting predicted.

I'm not sure if I'll add a minimum node size, I've left it up to my advisor. Personally, I think the point of random forest is allowing for the maximum possible reduction of Gini Impurity in every tree and then fixing the resulting overfitting using the forest's mean, which, to me, means intentionally preventing the Gini from reaching the minimum goes against the spirit of the algo. It *is* a valid solution though so who knows.